ntripbrowser: introduce new DataFetcher class
This class frees up `NtripBrowser` from work connected with data fetching.

URL's probing now executed in parallel, with use of MultiCurl.
In case of successful result on any of the input URL's it will execute a
`NtripBrowswer._process_raw_data()` function on it.

This approach allows to speed up data processing.